First Version of Doctor Java Class
Here is the first version of the Doctor Java Class, it contains 4 private variables including doctor Id, last name, first name and specialization. It also includes Getters for all these variables, a constructor that sets these variables and a "getAppointment" method that is not completely but will ultimately return all the rows in appointment that have the doctor's ID in them.